% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocColors.R
\name{biocColors}
\alias{biocColors}
\title{Colours for Bioconductor}
\usage{
biocColors(palette_name, n)
}
\arguments{
\item{palette_name}{The name of the palette to be returned.}

\item{n}{Integer which controls how many colors should be retrieved.}
}
\value{
A character vector of length \code{n}, comprising colors from
the palette specified by \code{palette_name}.
}
\description{
Colours for Bioconductor
}
\examples{
biocColors("ditto")

}
\author{
Daniel Bunis

Aaron Lun
}
